+++ bit/cli/mapped_reads_pid.py
@@ -1,7 +1,7 @@
 import sys
 import argparse
 from bit.cli.common import CustomRichHelpFormatter, add_help
-from bit.modules.get_mapped_reads_pid import (get_mapped_reads_pids,
+from bit.modules.mapped_reads_pid import (get_mapped_reads_pids,
                                               get_summary_stats)